Extract path rendering into PathRenderer
